modify the nodes
