+++ assets/reports/test.md
@@ -1,69 +1,1 @@
-# Morning Market Brief – April 30, 2025
-
-## Executive Summary
-- US equities closed higher on April 29, driven by robust corporate earnings.
-- Indian markets ended mixed, with sector-specific gains and losses amid geopolitical concerns.
-- Overnight, the US Dollar strengthened while commodity prices declined; gold prices held steady.
-- Key economic data releases today include Australia’s CPI, US GDP, and ADP employment report.
-- Microsoft and Volkswagen are notable companies reporting earnings.
-- Geopolitical developments include Russia-Ukraine ceasefire talks, widened US-Iran-China sanctions, and UN briefings on global hotspots.
-
-## Market Context
-US markets showed bullish momentum on strong earnings, easing concerns despite ongoing trade tensions. Indian markets reflected cautious investor stance influenced by geopolitical tensions and mixed earnings results. Overnight, global markets reacted to a stronger US Dollar and lower commodity prices, signaling subdued inflation expectations.
-
-## Key Market Highlights from April 29, 2025
-| Market/Index       | Close/Movement                         | Notes                        |
-|--------------------|---------------------------------------|------------------------------|
-| S&P 500 (US)       | Closed higher                         | Strong earnings reports      |
-| Dow Jones (US)     | Closed higher                         | Earnings-driven gains        |
-| Nasdaq (US)        | Closed higher                         | Tech sector contributed      |
-| Nifty 50 (India)   | 24,334.20, slight dip                 | Below 50 & 200 EMA           |
-| BSE Sensex (India) | Declined slightly (-0.06%)             | Mixed sector performance     |
-
-### Notable Sector Movements in India
-- Decliners: Bajaj Finserv (-6%), Bajaj Finance (-4%), Tata Motors, IndusInd Bank, SBI, UltraTech Cement
-- Gainers: Power Grid, HDFC Bank, Hindustan Unilever, NTPC, Mahindra & Mahindra, Maruti Suzuki
-- IT sector faced pressure from disappointing Q4 earnings of TCS, Infosys, HCL Tech, and Wipro
-
-## Overnight Global Activity and Currency Movements
-| Asset/Market      | Movement/Status                                    |
-|-------------------|--------------------------------------------------|
-| US Dollar (USD)   | Strengthened against major currencies            |
-| Gold              | Held steady above $3,300                          |
-| WTI Oil           | Declined to around $59.50                         |
-| Commodity Prices  | Broad-based decline aiding inflation control     |
-| Major Currencies  | Swiss franc and yen remained firm                 |
-
-Market sentiment remains cautious amid US-China trade frictions and global economic uncertainties. The stronger USD and softer commodities suggest muted near-term inflation pressures.
-
-## Key Economic Data Releases (April 30, 2025)
-| Time (ET) | Release                     | Significance                  |
-|-----------|-----------------------------|-------------------------------|
-| 4:30 am   | Australia CPI q/q            | Inflation indicator           |
-| 7:30 am   | US GDP (Q1)                 | Economic growth measure       |
-| 8:15 am   | US ADP Employment Report    | Labor market sentiment gauge  |
-
-## Corporate Earnings Announcements
-| Company     | Details                                  | Time (ET)         |
-|-------------|------------------------------------------|-------------------|
-| Volkswagen  | Q1 2025 Earnings Call                    | 3:00 am           |
-| Microsoft   | Q3 FY 2025 Earnings Call                 | 5:30 pm           |
-| Ansys       | Q1 2025 Earnings (post-market)           | After Market Close |
-| ACM Research| Q1 2025 Results & Revenue Outlook (recent)| April 29 release  |
-
-## Geopolitical Events
-- Russia awaits Ukraine’s response to a proposed May ceasefire.
-- US and Iran intensify sanctions on firms linked to China and Iran.
-- UN Security Council briefs on Libya, Haiti, and Colombia developments.
-
-These geopolitical tensions contribute to cautious investor behavior and potential market volatility.
-
-## Outlook and Implications
-- Today’s US macroeconomic data, particularly GDP and employment numbers, are key for assessing economic momentum and guiding market direction.
-- Strong earnings from Microsoft and Volkswagen may lift technology and automotive sectors.
-- Continued geopolitical risks and currency strength warrant monitoring for volatility in both Indian and global markets.
-- The decline in commodity prices may pressure resource-heavy sectors but help alleviate inflation concerns.
-
----
-
-This brief synthesizes market performance, overnight global developments, critical economic releases, earnings updates, and geopolitical factors as of April 30, 2025, to prepare investors for the day ahead.
+# No final report generated
