--- src/agent/market_intelligence_agent/tools/mcp_server_research.py
@@ -1,64 +1,0 @@
-import asyncio
-from langchain_mcp_adapters.client import MultiServerMCPClient
-from typing import List, Any, Dict, Callable
-from pathlib import Path
-
-base_path = Path(__file__).parent.parent.parent.parent.parent
-
-# Configuration for the MCP servers
-MCP_SERVERS_RESEARCH = {
-    "tavily_search": {
-        "command": "python",
-        "args": [str(base_path / "src/mcp_server/tavily.py")],
-        "transport": "stdio",
-    },
-    "tickertick": {
-        "command": "python",
-        "args": [str(base_path / "src/mcp_server/tickertick.py")],
-        "transport": "stdio",
-    }
-}
-
-async def execute_with_research_tools(operation: Callable[[List[Any], Dict[str, Any]], Any], state: Dict[str, Any]):
-    """
-    Creates a new MCP client, connects to research servers, performs the operation, 
-    and ensures proper cleanup within the same task.
-    
-    Args:
-        operation: A callable that takes (tools, state) and returns a result
-        state: The state dictionary to pass to the operation
-        
-    Returns:
-        The result of the operation
-    """
-    # Create a new client for this specific operation
-    client = MultiServerMCPClient()
-    
-    # Setup phase
-    try:
-        # Start the client
-        await client.__aenter__()
-        
-        # Connect to all servers
-        connect_tasks = []
-        for name, params in MCP_SERVERS_RESEARCH.items():
-            connect_tasks.append(client.connect_to_server(name, **params))
-        
-        await asyncio.gather(*connect_tasks)
-        
-        # Get tools
-        tools = client.get_tools()
-        
-        # Perform the operation
-        result = await operation(tools, state)
-        
-        return result
-    
-    finally:
-        # Always clean up
-        try:
-            await client.__aexit__(None, None, None)
-        except Exception as e:
-            print(f"Warning: Error during research client cleanup: {e}")
-            # Continue execution even if cleanup fails
-
